generated/cfhd_12.5_25_50-local/low-latency-initialization__t2.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Low-Latency (1): Initialization</title>
    <link rel="stylesheet" href="../lib/style.css" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>
    <div id="content-wrapper">
      <div id="video-wrapper">
        <video></video>
        <div><div id="qr-code"></div></div>
        <div><div id="status-text"></div></div>
      </div>
      <div>To show/hide debug overlay press up/down or button beneath</div>
      <div id="debug-wrapper">
        <div>
          <div class="button" id="debug-button">Debug Overlay</div>
        </div>
        <div id="debug"></div>
        <div id="log"></div>
      </div>
    </div>
    <div id="info-overlay"></div>
    <script src="../lib/mozilla/object-keys-polyfill.js"></script>
    <script src="../lib/stefanpenner/es6-promise.min.js"></script>
    <script src="../lib/player.js"></script>
    <script src="../lib/manifest_parser.js"></script>
    <script src="../lib/mpd-parser.js"></script>
    <script src="../lib/wave-service.js"></script>
    <script src="../lib/davidshimjs/qrcode.js"></script>
    <script src="../lib/dpctf-testharness.js"></script>
    <script>
      // Global variables
      var TEST_INFO = {
        id: "low-latency-initialization",
        title: "Low-Latency (1): Initialization",
        description: "",
        code: "low-latency-initialization.html",
      };

      var video = document.querySelector("video");
      var qrCode = document.getElementById("qr-code");
      var statusText = document.getElementById("status-text");

      var videoContentModel = ["/content/cfhd_sets/12.5_25_50/t2/2023-04-28/stream.mpd"];
      var audioContentModel = [];

      var dpctfTest = new DpctfTest({
        testInfo: TEST_INFO,
        videoContentModel: videoContentModel,
        audioContentModel: audioContentModel,
        videoElement: video,
        qrCodeElement: qrCode,
        statusTextElement: statusText,
        infoOverlayElement: document.getElementById("info-overlay"),
        executeTest: executeTest,
        usePlayout: false,
        initCallback: handleVideoInitialized,
      });

      function handleVideoInitialized(player, done, parameters) {
        var READY_STATE_HAVE_METADATA = 1;
        var waitingTime = parameters.waitingTime || 500;
        var handlePlay = function () {
          video.removeEventListener("play", handlePlay);
          dpctfTest.test(function () {
            assert_true(video.readyState === READY_STATE_HAVE_METADATA);
          }, "Video ready state is 'HAVE_METADATA'");
          setTimeout(function () {
            done();
            dpctfTest.setVideoState("appended");
          }, waitingTime);
        };
        video.addEventListener("play", handlePlay);
        player.play();
      }

      function executeTest(player, done, parameters) {
        var minBufferDuration = parameters.minBufferDuration / 1000;
        player.startBuffering();

        dpctfTest.asyncTest(function (test) {
          var query = location.search.replace(/\?/, "");
          var match = query.match(/redirect_time=([^&]+)/);
          var REDIRECT_TIME = match ? match[1] : null;
          if (!REDIRECT_TIME) REDIRECT_TIME = 5;
          video.addEventListener("ended", function (event) {
            assert_true(true);
            test.done();
          });
        }, "video ended event fired");

        done();
      }
    </script>
  </body>
</html>
